outsidehacks/requirements.txt
Django==1.8.3
Markdown==2.6.2
Pillow==2.9.0
django-filter==0.10.0
djangorestframework==3.1.3
psycopg2==2.6.1
wsgiref==0.1.2
